pyproject.toml
[project]
name = "rho-agent"
version = "0.1.0"
description = "An agent harness and CLI with readonly and developer modes."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.14.0",
    "prompt-toolkit>=3.0.52",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.0",
    "jinja2>=3.1.6",
]

[project.scripts]
rho-agent = "rho_agent.cli:cli"
rho-eval = "rho_agent.eval.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
# Database drivers
db-postgres = ["psycopg[binary]>=3.3.2"]
db-mysql = ["mysql-connector-python>=8.0"]
db-oracle = ["oracledb>=3.4.1"]
db-vertica = ["vertica-python>=1.4.0"]
db = ["rho-agent[db-postgres,db-mysql,db-oracle,db-vertica]"]

# Dashboard (Streamlit + pandas)
dashboard = ["streamlit>=1.40", "pandas>=2.0"]

# Excel file support
excel = ["openpyxl>=3.1.0"]

# Harbor integration for TerminalBench evaluation
# Install harbor from: https://github.com/laude-institute/harbor
harbor = []  # Harbor is installed separately via pip install -e ~/proj/harbor

# Eval dependencies (LiteLLM for multi-provider support and cost tracking)
evals = ["litellm>=1.50.0"]

# Observability dependencies
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]

# Observability: PostgreSQL backend (multi-user, multi-node)
obs-postgres = ["psycopg[binary,pool]>=3.3.2"]  # adds connection pooling on top of db-postgres

# Optional: OTLP export to external backends (Jaeger, Datadog, etc.)
otlp = [
    "opentelemetry-exporter-otlp>=1.20.0",
]

# Optional: Redis for real-time session tracking
redis = [
    "redis>=5.0.0",
]

# Optional: Daytona cloud sandbox integration
daytona = ["daytona>=0.1.0"]

# Everything
all = ["rho-agent[db,dashboard,excel,daytona,observability,obs-postgres,otlp,redis,evals]"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
]

[tool.ruff.lint]
select = ["E9", "F63", "F7", "F82"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
